Fixed a bug with RouletteSelector when the total fitness is zero. Removed an unneeded null check from TournamentSelector.

+++ src/main/java/com/ciphertool/genetics/algorithms/selection/modes/RouletteSelector.java
@@ -43,7 +43,7 @@ public class RouletteSelector implements Selector {
 			return -1;
 		}
 
-		long randomIndex = (int) (ThreadLocalRandom.current().nextDouble() * totalFitness);
+		double randomIndex = (ThreadLocalRandom.current().nextDouble() * totalFitness);
 
 		int winningIndex = -1;
 		Chromosome nextIndividual = null;
